Clippy

+++ examples/canvas_subframe.rs
@@ -88,10 +88,10 @@ impl event::EventHandler<ggez::GameError> for MainState {
 
         // Bounce the rect if necessary
         let (w, h) = ctx.gfx.drawable_size();
-        if self.draw_pt.x + (w as f32 / 2.0) > (w as f32) || self.draw_pt.x < 0.0 {
+        if self.draw_pt.x + w / 2.0 > w || self.draw_pt.x < 0.0 {
             self.draw_vec.x *= -1.0;
         }
-        if self.draw_pt.y + (h as f32 / 2.0) > (h as f32) || self.draw_pt.y < 0.0 {
+        if self.draw_pt.y + h / 2.0 > h || self.draw_pt.y < 0.0 {
             self.draw_vec.y *= -1.0;
         }
         self.draw_pt += self.draw_vec;
