+++ examples/graphics_settings.rs
@@ -157,16 +157,14 @@ impl event::EventHandler<ggez::GameError> for MainState {
                 self.zoom += 0.1;
                 println!("Zoom is now {}", self.zoom);
                 let (w, h) = ctx.gfx.drawable_size();
-                let new_rect =
-                    graphics::Rect::new(0.0, 0.0, w as f32 * self.zoom, h as f32 * self.zoom);
+                let new_rect = graphics::Rect::new(0.0, 0.0, w * self.zoom, h * self.zoom);
                 self.screen_coords = new_rect;
             }
             Some(KeyCode::Down) => {
                 self.zoom -= 0.1;
                 println!("Zoom is now {}", self.zoom);
                 let (w, h) = ctx.gfx.drawable_size();
-                let new_rect =
-                    graphics::Rect::new(0.0, 0.0, w as f32 * self.zoom, h as f32 * self.zoom);
+                let new_rect = graphics::Rect::new(0.0, 0.0, w * self.zoom, h * self.zoom);
                 self.screen_coords = new_rect;
             }
             Some(KeyCode::Space) => {
@@ -184,12 +182,7 @@ impl event::EventHandler<ggez::GameError> for MainState {
     fn resize_event(&mut self, _ctx: &mut Context, width: f32, height: f32) -> GameResult {
         println!("Resized screen to {}, {}", width, height);
         if self.window_settings.resize_projection {
-            let new_rect = graphics::Rect::new(
-                0.0,
-                0.0,
-                width as f32 * self.zoom,
-                height as f32 * self.zoom,
-            );
+            let new_rect = graphics::Rect::new(0.0, 0.0, width * self.zoom, height * self.zoom);
             self.screen_coords = new_rect;
         }
         Ok(())
